refactor: 사용자 신고와 관련된 로직 위치 수정
