--- src/main/java/com/kgu/studywithme/member/service/MemberReportService.java
@@ -1,32 +1,0 @@
-package com.kgu.studywithme.member.service;
-
-import com.kgu.studywithme.global.exception.StudyWithMeException;
-import com.kgu.studywithme.member.domain.MemberRepository;
-import com.kgu.studywithme.member.domain.report.Report;
-import com.kgu.studywithme.member.domain.report.ReportRepository;
-import com.kgu.studywithme.member.exception.MemberErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class MemberReportService {
-    private final ReportRepository reportRepository;
-    private final MemberRepository memberRepository;
-
-    @Transactional
-    public Long report(Long reporteeId, Long reporterId, String reason) {
-        validatePreviousReportIsStillPending(reporteeId, reporterId);
-
-        Report report = Report.createReportWithReason(reporteeId, reporterId, reason);
-        return reportRepository.save(report).getId();
-    }
-
-    private void validatePreviousReportIsStillPending(Long reporteeId, Long reporterId) {
-        if (memberRepository.isReportReceived(reporteeId, reporterId)) {
-            throw StudyWithMeException.type(MemberErrorCode.REPORT_IS_STILL_RECEIVED);
-        }
-    }
-}
