+++ src/main/java/com/kgu/studywithme/member/service/MemberService.java
@@ -1,7 +1,11 @@
 package com.kgu.studywithme.member.service;
 
+import com.kgu.studywithme.global.exception.StudyWithMeException;
 import com.kgu.studywithme.member.domain.Member;
 import com.kgu.studywithme.member.domain.MemberRepository;
+import com.kgu.studywithme.member.domain.report.Report;
+import com.kgu.studywithme.member.domain.report.ReportRepository;
+import com.kgu.studywithme.member.exception.MemberErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberService {
     private final MemberValidator memberValidator;
     private final MemberRepository memberRepository;
+    private final ReportRepository reportRepository;
 
     @Transactional
     public Long signUp(Member member) {
@@ -23,5 +28,19 @@ public class MemberService {
         memberValidator.validateEmail(member.getEmail());
         memberValidator.validateNickname(member.getNickname());
         memberValidator.validatePhone(member.getPhone());
+    }
+
+    @Transactional
+    public Long report(Long reporteeId, Long reporterId, String reason) {
+        validatePreviousReportIsStillPending(reporteeId, reporterId);
+
+        Report report = Report.createReportWithReason(reporteeId, reporterId, reason);
+        return reportRepository.save(report).getId();
+    }
+
+    private void validatePreviousReportIsStillPending(Long reporteeId, Long reporterId) {
+        if (memberRepository.isReportReceived(reporteeId, reporterId)) {
+            throw StudyWithMeException.type(MemberErrorCode.REPORT_IS_STILL_RECEIVED);
+        }
     }
 }
