+++ src/test/java/com/kgu/studywithme/common/ControllerTest.java
@@ -15,10 +15,8 @@ import com.kgu.studywithme.favorite.controller.FavoriteApiController;
 import com.kgu.studywithme.favorite.service.FavoriteManageService;
 import com.kgu.studywithme.member.controller.MemberApiController;
 import com.kgu.studywithme.member.controller.MemberInformationApiController;
-import com.kgu.studywithme.member.controller.MemberReportApiController;
 import com.kgu.studywithme.member.service.MemberFindService;
 import com.kgu.studywithme.member.service.MemberInformationService;
-import com.kgu.studywithme.member.service.MemberReportService;
 import com.kgu.studywithme.member.service.MemberService;
 import com.kgu.studywithme.study.controller.*;
 import com.kgu.studywithme.study.controller.notice.StudyNoticeApiController;
@@ -53,7 +51,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
         CategoryApiController.class, StudyParticipationApiController.class, StudyApiController.class,
         FavoriteApiController.class, StudySearchApiController.class, StudyInformationApiController.class,
         StudyNoticeApiController.class, MemberInformationApiController.class, StudyNoticeCommentApiController.class,
-        StudyReviewApiController.class, MemberReportApiController.class
+        StudyReviewApiController.class
 })
 @ExtendWith(RestDocumentationExtension.class)
 @Import(TestAopConfiguration.class)
@@ -115,9 +113,6 @@ public abstract class ControllerTest {
 
     @MockBean
     protected StudyReviewService studyReviewService;
-
-    @MockBean
-    protected MemberReportService memberReportService;
 
     @BeforeEach
     void setUp(WebApplicationContext context, RestDocumentationContextProvider provider) {
